+++ jasmine.json
@@ -1,7 +1,7 @@
 {
   "spec_dir": "test",
   "spec_files": [
-    "**/AllOutAssault.spec.js"
+    "**/*[sS]pec.js"
   ],
   "helpers": [
     "helpers/**/*.js"
